factor transaction out to a trait
