+++ project/build.properties
@@ -4,5 +4,5 @@ project.organization=com.twitter
 project.name=querulous
 sbt.version=0.7.4
 project.version=1.5.5-SNAPSHOT
-build.scala.versions=2.7.7 2.8.0
+build.scala.versions=2.7.7 2.8.0 2.8.1
 project.initialize=false
